% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkerServer.R
\name{checkerServer}
\alias{checkerServer}
\title{Data Quality Check Server}
\usage{
checkerServer(
  id,
  date_var,
  partner_var,
  state_var,
  lga_var,
  facility_var,
  var_names
)
}
\arguments{
\item{id}{Namespace id for the Shiny module}

\item{date_var}{the variable in the data set to be used for the calendar date}

\item{partner_var}{the 'Implementing Partner' variable from the data set}

\item{state_var}{the reference variable for 'State' in the data set}

\item{lga_var}{the reference variable for 'LGA' in the data set}

\item{facility_var}{the reference variable for 'Facility' in the data set}

\item{var_names}{all the variable names to be extracted from the supplied data set}
}
\value{
server component
}
\description{
Data Quality Check Server
}
\examples{
NULL
}
\keyword{internal}
