% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkerUI.R
\name{checkerUI}
\alias{checkerUI}
\title{User Interface}
\usage{
checkerUI(id, .opts, .mult)
}
\arguments{
\item{id}{Namespace id for the Shiny module}

\item{.opts}{Download options}

\item{.mult}{Logical, whether user can select multiple download options at a time}
}
\value{
user interface
}
\description{
User Interface
}
\examples{
NULL

}
\keyword{internal}
