% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_theme.R
\name{default_theme}
\alias{default_theme}
\title{Default Theme}
\usage{
default_theme()
}
\value{
no return value
}
\description{
Set default theme for the Shiny applications
}
\examples{
NULL
}
\keyword{internal}
