% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recency_dqa.R
\name{dqa_pos}
\alias{dqa_pos}
\title{Perform Recency Data Quality Check}
\usage{
dqa_pos(df)
}
\arguments{
\item{df}{dataframe of recency line-list}
}
\value{
table of data quality check
}
\description{
Perform Recency Data Quality Check
}
\keyword{internal}
