% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checkndr.R
\name{hts_download_opts}
\alias{hts_download_opts}
\title{HTS Download Options}
\usage{
hts_download_opts()
}
\value{
NULL. Used for side effect
}
\description{
HTS Download Options
}
\examples{
NULL

}
\keyword{internal}
