% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checkndr.R
\name{hts_var_names}
\alias{hts_var_names}
\title{HTS Line-list Variables}
\usage{
hts_var_names()
}
\value{
NULL. Used for side effect
}
\description{
HTS Line-list Variables
}
\examples{
NULL
}
\keyword{internal}
