% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htsApp.R
\name{htsApp}
\alias{htsApp}
\title{Start the HTS Data Quality Checker App}
\usage{
htsApp(.theme = NULL)
}
\arguments{
\item{.theme}{customise the appearance of your Shiny app. This can be done using the `bslib::bs_theme()` function.}
}
\value{
HTS data quality app
}
\description{
Start the HTS Data Quality Checker App
}
\examples{
## Only run examples in interactive R sessions
if (interactive()) {
  htsApp()
}

}
\keyword{internal}
