% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_file.R
\name{load_file}
\alias{load_file}
\title{Upload Line List}
\usage{
load_file(name, path, var_names)
}
\arguments{
\item{name}{name of line-list file}

\item{path}{location of file on the computer}

\item{var_names}{variable names of the line-list}
}
\value{
dataframe of line-list
}
\description{
Upload Line List
}
\examples{
NULL
}
\keyword{internal}
