% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickInput.R
\name{pickInput}
\alias{pickInput}
\title{Title Select Input}
\usage{
pickInput(id, lab, opts = NULL, mult = TRUE)
}
\arguments{
\item{id}{input ID}

\item{lab}{title of the select input}

\item{opts}{select options}

\item{mult}{boolean, should multiple selections be possible?}
}
\value{
selected input
}
\description{
Title Select Input
}
\examples{
NULL
}
\keyword{internal}
