% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_recency.R
\name{recency_download_opts}
\alias{recency_download_opts}
\title{Recency Download Options}
\usage{
recency_download_opts()
}
\value{
NULL. Used for side effect
}
\description{
Recency Download Options
}
\examples{
NULL

}
\keyword{internal}
