% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recency_dqa.R
\name{recency_dqa}
\alias{recency_dqa}
\title{Perform Recency Data Quality Check}
\usage{
recency_dqa(df)
}
\arguments{
\item{df}{dataframe of recency line-list}
}
\value{
table of data quality check
}
\description{
Perform Recency Data Quality Check
}
\examples{
NULL
}
\keyword{internal}
