% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checkndr.R
\name{recency_var_names}
\alias{recency_var_names}
\title{Recency Line-list Variables}
\usage{
recency_var_names()
}
\value{
NULL. Used for side effect
}
\description{
Recency Line-list Variables
}
\examples{
NULL
}
\keyword{internal}
