% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recencyApp.R
\name{recencyApp}
\alias{recencyApp}
\title{Start the Recency Data Quality Checker App}
\usage{
recencyApp(.theme = NULL)
}
\arguments{
\item{.theme}{customise the appearance of your Shiny app. This can be done using the `bslib::bs_theme()` function.}
}
\value{
recency data quality app
}
\description{
Start the Recency Data Quality Checker App
}
\examples{
## Only run examples in interactive R sessions
if (interactive()) {
  recencyApp()
}

}
