% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htsApp.R, R/recencyApp.R
\name{table_dqa}
\alias{table_dqa}
\title{Specify the Function for the Data Quality}
\usage{
table_dqa(df)

table_dqa(df)
}
\arguments{
\item{df}{dataframe of line list}
}
\value{
table of data quality check outcome

table of data quality check outcome
}
\description{
Specify the Function for the Data Quality

Specify the Function for the Data Quality
}
\examples{
NULL
NULL
}
\keyword{internal}
