% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recencyApp.R
\name{table_dqa_pos}
\alias{table_dqa_pos}
\title{Specify the Function for the Data Quality}
\usage{
table_dqa_pos(df)
}
\arguments{
\item{df}{dataframe of line list}
}
\value{
table of data quality check outcome
}
\description{
Specify the Function for the Data Quality
}
\examples{
NULL
}
\keyword{internal}
