% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_cases.R
\name{valid_cases}
\alias{valid_cases}
\title{Check for Number of rows with Valid Entries}
\usage{
valid_cases(df)
}
\arguments{
\item{df}{dataframe of line list}
}
\value{
table of rows with valid entries in all cells
}
\description{
Check for Number of rows with Valid Entries
}
\examples{
NULL
#  df |>
#    subset(
#      !is.na(client_state) &
#      !is.na(client_lga) &
#      sex \%in\% c("Male", "Female", "M", "F", "male", "female", "m", "f") &
#      age >= 15 &
#      !is.na(visit_date) &
#      hts_result \%in\% c("R", "Pos") &
#      hts_confirmatory_result \%in\% c("R", "Pos") | hts_tie_breaker_result \%in\% c("R", "Pos") &
#      !is.na(opt_out) &
#      !is.na(testing_point) &
#      recency_test_name \%in\% c("Asante", "AS") &
#      !is.na(recency_test_date) &
#      !is.na(recency_number) &
#      control_line \%in\% c("Yes", "No") &
#      verification_line \%in\% c("Yes", "No") &
#      longterm_line \%in\% c("Yes", "No") &
#      control_line \%in\% "Yes" & verification_line \%in\% "Yes" & longterm_line \%in\% "Yes" & recency_interpretation \%in\% "LongTerm" |
#      control_line \%in\% "Yes" & verification_line \%in\% "Yes" & longterm_line \%in\% "No" & recency_interpretation \%in\% "Recent" |
#      control_line \%in\% "Yes" & verification_line \%in\% "No" & longterm_line \%in\% "No" & recency_interpretation \%in\% "Negative" |
#      control_line \%in\% "No" & recency_interpretation \%in\% "Invalid" |
#      control_line \%in\% "Yes" & verification_line \%in\% "No" & longterm_line \%in\% "Yes" & recency_interpretation \%in\% "Invalid"|
#      control_line \%in\% "Yes" & verification_line \%in\% "Yes" & longterm_line \%in\% "No" & viral_load_requested \%in\% "Yes" &
#      !is.na(date_sample_collected) &
#      ifelse(
#        recency_interpretation \%in\% "Recent" & visit_date + lubridate::days(42) < max(visit_date, na.rm = TRUE),
#        !is.na(date_of_viral_load_result),
#        is.na(date_of_viral_load_result)
#      )
#    )
}
\keyword{internal}
