% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_results.R
\name{write_results}
\alias{write_results}
\title{Write Multiple Excel Worksheets}
\usage{
write_results(list_data, f_name, ...)
}
\arguments{
\item{list_data}{datasets to be written into worksheets}

\item{f_name}{name of Excel workbook}

\item{...}{other arguments supplied to `openxlsx::writeData()`}
}
\value{
Excel workbook
}
\description{
Write Multiple Excel Worksheets
}
\examples{
NULL
}
\keyword{internal}
