% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkndr-package.R
\docType{package}
\name{checkndr-package}
\alias{checkndr}
\alias{checkndr-package}
\title{checkndr: Data Quality Checks for the Nigeria National Data Repository (NDR) Patient-level Line Lists}
\description{
Performs data quality checks using the standard quality metrics on patient-level line-lists downloaded from the NDR or similar platforms. Currenlty, it supports the Recency and Mortality Surveillance line-lists.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/stephenbalogun/checkndr}
  \item Report bugs at \url{https://github.com/stephenbalogun/checkndr/issues}
}

}
\author{
\strong{Maintainer}: Stephen Balogun \email{stephentaiyebalogun@gmail.com} (\href{https://orcid.org/0000-0002-9928-3703}{ORCID})

Other contributors:
\itemize{
  \item Kehinde Balogun [contributor]
}

}
\keyword{internal}
