% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recencyApp.R
\name{filter_valid}
\alias{filter_valid}
\title{Filter rows with Complete Valid Entries}
\usage{
filter_valid(df)
}
\arguments{
\item{df}{dataframe of line list}
}
\value{
table of rows with valid entries in all cells
}
\description{
Filter rows with Complete Valid Entries
}
\examples{
NULL
}
\keyword{internal}
