% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recency_line_list.R
\name{recency_line_list}
\alias{recency_line_list}
\title{Generate Line-list of Selected Indicators}
\usage{
recency_line_list(df, input)
}
\arguments{
\item{df}{Line-list supplied}

\item{input}{Selected indicator(s)}
}
\value{
dataframe
}
\description{
Generate Line-list of Selected Indicators
}
\examples{
NULL
}
\keyword{internal}
