% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recencyApp.R
\name{request_line_list}
\alias{request_line_list}
\title{Specify the Type of Line-list to be Requested}
\usage{
request_line_list(input, df)
}
\arguments{
\item{input}{Selected indicator(s)}

\item{df}{Line-list supplied}
}
\value{
line-list of requested indicator(s)
}
\description{
Specify the Type of Line-list to be Requested
}
\examples{
NULL
}
\keyword{internal}
