% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_cases.R
\name{valid_cases}
\alias{valid_cases}
\title{Check for Number of rows with Valid Entries}
\usage{
valid_cases(df)
}
\arguments{
\item{df}{dataframe of line list}
}
\value{
table of rows with valid entries in all cells
}
\description{
Check for Number of rows with Valid Entries
}
\examples{
NULL
}
\keyword{internal}
